Add 'Slate' style for 'Cover' blocks
A new 'Slate' style has been added for 'Cover' blocks in the block-styles.php file. This change is made visible in the hero.php file. This new style modifies the block appearance with new CSS rules defined in _cover.scss and connected with the main styles in the patterns.scss file. The new style enhances the visual contrast and makes the 'Cover' blocks more engaging.